refactor(backend): move auth module from core to services and enforce module boundaries with tach
Add tach.toml defining layered dependency rules (api → services → core/crud → db/schemas/settings), wire tach into the Makefile ci target and lefthook pre-commit hook, and update all import sites to reflect the core/auth → services/auth rename.
