+++ docs/src/partials/root_commands.md
@@ -6,7 +6,7 @@ make lint          # backend + frontend lint
 make fix           # autofix both
 make typecheck     # mypy + tsc
 make test          # backend pytest w/ coverage
-make ci            # lint + typecheck + test + docs-check
+make ci            # lint + typecheck + boundaries + test + docs-check
 make docs-build    # render docs/src → README.md, CLAUDE.md, AGENTS.md, sub-READMEs
 make docs-check    # fail if generated docs are stale
 make backend-<X>   # delegates to fastapi/Makefile target X (e.g. backend-logs, backend-health, backend-lint)
