+++ lefthook.yml
@@ -2,16 +2,26 @@ pre-commit:
   parallel: true
   commands:
     biome:
-      glob: "frontend/src/**/*.{ts,tsx,js,jsx,json}"
-      run: cd frontend && npx biome check --write --no-errors-on-unmatched {staged_files}
+      root: frontend/
+      glob: "src/**/*.{ts,tsx,js,jsx,json}"
+      run: npx biome check --write --no-errors-on-unmatched {staged_files}
       stage_fixed: true
 
     ruff-lint:
-      glob: "fastapi/**/*.py"
-      run: cd fastapi && uv run ruff check --fix {staged_files}
+      root: fastapi/
+      glob: "**/*.py"
+      run: uv run ruff check --fix {staged_files}
       stage_fixed: true
 
     ruff-format:
-      glob: "fastapi/**/*.py"
-      run: cd fastapi && uv run ruff format {staged_files}
+      root: fastapi/
+      glob: "**/*.py"
+      run: uv run ruff format {staged_files}
       stage_fixed: true
+
+    tach:
+      root: fastapi/
+      glob:
+        - "app/**/*.py"
+        - "tach.toml"
+      run: uv run tach check
